Privacy-Guard: Access Sensor-Template for Platform apps
Use of Motion Sensors is set to "ask" by default. Hence, also
platform apps ask for it. Platform apps however are - with the
exception of root - excluded from the Privacy Guard details screen.
In order to avoid the obvious conflict here, access to the sensor
template settings will also be allowed for platform apps.

Change-Id: I2408f212d27b45f169c56d242ced19d17bda23a3